--- MANIFEST.in
@@ -1,1 +1,0 @@
-global-exclude test_*.py
