+++ requirements.txt
@@ -1,6 +1,6 @@
 pandas==0.25.0
 scipy==1.2.2
-sklearn
+sklearn==0.21.3
 seaborn==0.9.0
 matplotlib==3.1.0
-imblearn
+imblearn==0.5.0
