+++ setup.py
@@ -13,7 +13,7 @@ setuptools.setup(
     long_description=long_description,
     long_description_content_type='text/markdown',
     url='https://github.com/resuelve/silk-ml',
-    packages=setuptools.find_packages(exclude=['sphinx_docs', 'docs']),
+    packages=setuptools.find_packages(exclude=['sphinx_docs', 'docs', 'tests']),
     python_requires='~=3.5',
     install_requires=[
         i.replace('\n', '')
@@ -27,7 +27,6 @@ setuptools.setup(
         'Intended Audience :: Science/Research',
         'Intended Audience :: Developers',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
         'Topic :: Software Development',
